requirements.txt
matplotlib
numpy 
seaborn
pillow
tensorflow
keras
gradio
opencv-pythonen